extract homepage component
